# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Todo Application REST-API scheme",
        "title": "Golang Todo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:5050",
    "basePath": "/api/v1",
    "paths": {
...
        "/tasks": {
            "get": {
                "description": "Список задач с опциональной пагинацией и/или фильтрацией по ID автора задачи",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Список задач",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Фильтрация задач по ID автора",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Размер страниц с задачами",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение страниц с задачами",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список задач",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_features_tasks_transport_http.TaskDTOResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_KirillSerge_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_KirillSerge_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создать новую заадачу в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Создать задачу",
                "parameters": [
                    {
                        "description": "CreateTask тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.CreateTaksRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешно созданная задача",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.CreateTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_KirillSerge_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Author not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_KirillSerge_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_KirillSerge_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "description": "Получение конкретной задачи по ее ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Получение задачи",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID получаемой задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задача успешно найдена",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.CreateTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_KirillSerge_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_KirillSerge_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_KirillSerge_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удалить существующию задачу по ID",
                "tags": [
                    "tasks"
                ],
                "summary": "Удалить задачу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID удаляемой задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Успешное удаление задачи"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_KirillSerge_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_KirillSerge_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_KirillSerge_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Обновляет информацию об уже существующей задаче\nИзменение информации об уже существующем в системе пользователи\n### Логика обновления полей (Three-state logic):\n1. **Поле не переданно**:`description` игнорируется, значение в БД не меняется\n2. **Явно передано значение**: `\"description\":\"выйти на прогулку с бобиком\"` - устанавливает новый номер телефона в БД\n3. **Передан null**: `\"description\":null` - очищает поле в БД (set to NULL)\nОграничение: `title`и `completed` не может быть выставлен как null",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Обновить задачу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID изменяемой задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PatchTask тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.PathTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Усрешно измененная задача",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_features_tasks_transport_http.TaskDTOResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_KirillSerge_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_KirillSerge_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_KirillSerge_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_KirillSerge_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_KirillSerge_golang-todoapp_internal_core_transport_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "full error text"
                },
                "message": {
                    "type": "string",
                    "example": "short human-readable message"
                }
            }
        },
        "github_com_KirillSerge_golang-todoapp_internal_core_transport_http_types.Nullable-bool": {
            "type": "object",
            "properties": {
                "set": {
                    "type": "boolean"
                },
                "value": {
                    "type": "boolean"
                }
            }
        },
        "github_com_KirillSerge_golang-todoapp_internal_core_transport_http_types.Nullable-string": {
            "type": "object",
            "properties": {
                "set": {
                    "type": "boolean"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "internal_features_tasks_transport_http.CreateTaksRequest": {
            "type": "object",
            "required": [
                "author_user_id",
                "title"
            ],
            "properties": {
                "author_user_id": {
                    "type": "integer",
                    "example": 5
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 1,
                    "example": "Сделать до четверга дз"
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "Домашнее задание"
                }
            }
        },
        "internal_features_tasks_transport_http.CreateTaskResponse": {
            "type": "object",
            "properties": {
                "author_user_id": {
                    "type": "integer",
                    "example": 5
                },
                "completed": {
                    "type": "boolean",
                    "example": false
                },
                "completed_at": {
                    "type": "string",
                    "example": "null"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-02-26T10:30:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Сделать до четверга"
                },
                "id": {
                    "type": "integer",
                    "example": 15
                },
                "title": {
                    "type": "string",
                    "example": "Домашка"
                },
                "version": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "internal_features_tasks_transport_http.PathTaskRequest": {
            "type": "object",
            "properties": {
                "completed": {
                    "$ref": "#/definitions/github_com_KirillSerge_golang-todoapp_internal_core_transport_http_types.Nullable-bool"
                },
                "description": {
                    "$ref": "#/definitions/github_com_KirillSerge_golang-todoapp_internal_core_transport_http_types.Nullable-string"
                },
                "title": {
                    "$ref": "#/definitions/github_com_KirillSerge_golang-todoapp_internal_core_transport_http_types.Nullable-string"
                }
            }
        },
        "internal_features_tasks_transport_http.TaskDTOResponse": {
            "type": "object",
            "properties": {
                "author_user_id": {
                    "type": "integer",
                    "example": 5
                },
                "completed": {
                    "type": "boolean",
                    "example": false
                },
                "completed_at": {
                    "type": "string",
                    "example": "null"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-02-26T10:30:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Сделать до четверга"
                },
                "id": {
                    "type": "integer",
                    "example": 15
                },
                "title": {
                    "type": "string",
                    "example": "Домашка"
                },
                "version": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
    }
}